// components/searchSelect/targets/searchSelectTargets.enum.ts
export enum SearchSelectTargets {
  ROOT = 'root',
  INPUT = 'input',
  LABEL = 'label',
  SEARCH_INPUT = 'searchInput',
  CUSTOM_POPPER = 'customPopper',
  TAG = 'tag',
  READ_ONLY = 'readOnly',
  ERRORS_ROOT = 'errorsRoot',
  ERRORS_UNORDERED_LIST = 'errorsUnorderedList',
  ERRORS_UNORDERED_LIST_ITEM = 'errorsUnorderedListItem',
  ERRORS_FIELD_ERROR = 'errorsFieldError',
}